pyproject.toml
[tool.poetry]
name = "sanskrit-translator-dspy"
version = "0.1.0"
description = ""
authors = ["Ganaraj Permunda <ganarajpr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dspy-ai = "^2.4.9"
fastapi = "^0.111.0"
numpy = "^2.0.0"
ollama = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
